Refactor into Post Class
This separates all of the json handling into Post.from_json and
fetch_posts. The non-parallel version of download_all_images has been
removed because it was dead code. A Url phantom type is added to ensure
it's always clear whether a subreddit is a name or a url.

Two unit tests are added for the Post class with an valid post JSON and
a self-post JSON.